requirements.txt
absl-py==2.3.1
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
importlib_metadata==8.7.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.0.2
opt_einsum==3.4.0
optree==0.18.0
outcome==1.3.0.post0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.33.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
selenium==4.36.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wrapt==2.0.1
wsproto==1.2.0
yarg==0.1.9
zipp==3.23.0
